requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
pytest-shutil